[#92] Move header files to centralised dir
Finished #92
