+++ inc/Coin.h
