+++ inc/Global.h
